JCBC-1645 Followup: fix KeyValueIntegrationTest.checkExpiryBeyond2038
Motivation
----------
Couchbase Server interprets expiry epoch seconds as unsigned integers,
meaning that we have some headroom well beyond 2038.

Modifications
-------------
Don't cap the expiry epoch second at 2038.

Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/132041
Tested-by: Build Bot <build@couchbase.com>
(cherry picked from commit 19a583989717d781b20cd99ccc974f82184c3b90)
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-clients/+/131954

+++ java-client/src/main/java/com/couchbase/client/java/AsyncBinaryCollection.java
@@ -205,7 +205,7 @@ public class AsyncBinaryCollection {
     RetryStrategy retryStrategy = opts.retryStrategy().orElse(environment.retryStrategy());
     InternalSpan span = environment.requestTracer().internalSpan(IncrementRequest.OPERATION_NAME, opts.parentSpan().orElse(null));
 
-    int expiry = ExpiryUtils.getAdjustedExpirySeconds(opts.expiry(), environment.eventBus());
+    long expiry = ExpiryUtils.getAdjustedExpirySeconds(opts.expiry(), environment.eventBus());
     IncrementRequest request = new IncrementRequest(timeout, coreContext, collectionIdentifier, retryStrategy, id, opts.cas(),
       opts.delta(), opts.initial(), expiry, opts.durabilityLevel(), span);
     request.context().clientContext(opts.clientContext());
@@ -249,7 +249,7 @@ public class AsyncBinaryCollection {
     RetryStrategy retryStrategy = opts.retryStrategy().orElse(environment.retryStrategy());
     InternalSpan span = environment.requestTracer().internalSpan(DecrementRequest.OPERATION_NAME, opts.parentSpan().orElse(null));
 
-    int expiry = ExpiryUtils.getAdjustedExpirySeconds(opts.expiry(), environment.eventBus());
+    long expiry = ExpiryUtils.getAdjustedExpirySeconds(opts.expiry(), environment.eventBus());
     DecrementRequest request = new DecrementRequest(timeout, coreContext, collectionIdentifier, retryStrategy, id, opts.cas(),
       opts.delta(), opts.initial(), expiry, opts.durabilityLevel(), span);
     request.context().clientContext(opts.clientContext());
